Code/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
convertapi==1.4.0
cryptography==36.0.2
cycler==0.11.0
Deprecated==1.2.13
Flask==2.1.1
Flask-Caching==1.10.1
gunicorn==20.1.0
fonttools==4.31.2
gcloud==0.18.3
google-api-core==2.7.1
google-api-python-client==2.44.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jwcrypto==1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauth2client==4.1.3
opencv-contrib-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
pygame==2.1.2
pyparsing==3.0.7
Pyrebase4==4.5.0
python-dateutil==2.8.2
python-jwt==3.3.2
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==4.8
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
